refactor(ap): enforce approve permission inside invoice form
